requirements.txt
anthropic>=0.49.0
aiohttp>=3.9.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
certifi
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
